+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Also include relevant environment variables in cache keys
 - More statistics in `cache stats`
+- Add `cache entry` subcommand for freeform cache usage
 
 ## [0.5.0] - 2025-12-16